Time: 411 ms (5.09%), Space: 48.9 MB (53.53%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.java
@@ -1,6 +1,6 @@
 class Solution {
     public int numSubarrayProductLessThanK(int[] nums, int k) {
-        int pro=1;
+        long pro=1;
         int count = 0;
         for(int i=0;i<nums.length;i++){
             pro=1;
